localtm: implement Python importer
This is an initial rewrite of the Perl importer in Python.  Note this
does not yet use the bulk() import functionality, where it should
because of potential speed issues.

This command should migrate to a ./manage.py command so that it can more
easily access the configured Pootle database instead of the hardcoded
sqlite support currently in the implementation.